+++ keyboards/nebula68/info.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "NEBULA65",
+    "keyboard_name": "NEBULA68",
     "url": "",
     "maintainer": "yiancar",
    "width": 17.25,
