+++ keyboards/nebula68/keymaps/via/readme.md
@@ -1,7 +1,7 @@
-# The default keymap for Nebula65. VIA support enabled.
+# The default keymap for Nebula68. VIA support enabled.
 
 ![Layer 0](https://i.imgur.com/dXyRwb1.png)
 
 ![Layer 1](https://i.imgur.com/4izTAFa.png)
 
-Default layer is normal ANSI 65%
+Default layer is normal ANSI 68%
